Allow linking with iOS app extension
This removes the call to [UIApplication sharedApplication] in
WAAppRouteHandler.m, since that method is off-limits to iOS app
extensions.

WAAppRouteHandler's designated initializer has been altered
to accept a rootViewController:(UIViewController *) parameter, which it
uses instead of pulling the rootViewController from the application's
UIWindow.

+ [WAAppRouter defaultRouter] has been changed to + [WAAppRouter
defaultRouterWithRootViewController:(UIViewController)]

With these changes, it is possible to link WAAppRouting into a shared
framework that is linked to both an application and an iOS app
extension.