Avoid calling sharedApplication if WA_APP_EXTENSION flag is set
This wraps the call to [UIApplication sharedApplication] in `#ifndef WA_APP_EXTENSION`, to allow compiling it in an "app extension safe" mode.

This allows using WAAppRouting in a shared framework that is linked to
an app extension, so long as the WA_APP_EXTENSION flag is set when
building WAAppRouting.